tests/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.14
markdown>=3.4
django-filter>=22.1
requests>=2.28
requests-mock>=1.10
pytest>=7.1
pytest-cov>= 4.0
pytest-django>=4.5
pytest-mock>=3.10
model-bakery>=1.8
django-mock-queries>=2.1